Add orthographic camera

+++ ParasiteEngine/Source/ParasiteEngine/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -10,17 +10,17 @@ namespace Parasite
 	{
 		switch (InType)
 		{
-		case Parasite::EShaderDataType::Float:	return GL_FLOAT;
-		case Parasite::EShaderDataType::Float2: return GL_FLOAT;
-		case Parasite::EShaderDataType::Float3:	return GL_FLOAT;
-		case Parasite::EShaderDataType::Float4: return GL_FLOAT;
-		case Parasite::EShaderDataType::Mat3:	return GL_FLOAT;
-		case Parasite::EShaderDataType::Mat4:	return GL_FLOAT;
-		case Parasite::EShaderDataType::Int:	return GL_INT;
-		case Parasite::EShaderDataType::Int2:	return GL_INT;
-		case Parasite::EShaderDataType::Int3:	return GL_INT;
-		case Parasite::EShaderDataType::Int4:	return GL_INT;
-		case Parasite::EShaderDataType::Bool:	return GL_BOOL;
+			case Parasite::EShaderDataType::Float:		return GL_FLOAT;
+			case Parasite::EShaderDataType::Float2:		return GL_FLOAT;
+			case Parasite::EShaderDataType::Float3:		return GL_FLOAT;
+			case Parasite::EShaderDataType::Float4:		return GL_FLOAT;
+			case Parasite::EShaderDataType::Mat3:		return GL_FLOAT;
+			case Parasite::EShaderDataType::Mat4:		return GL_FLOAT;
+			case Parasite::EShaderDataType::Int:		return GL_INT;
+			case Parasite::EShaderDataType::Int2:		return GL_INT;
+			case Parasite::EShaderDataType::Int3:		return GL_INT;
+			case Parasite::EShaderDataType::Int4:		return GL_INT;
+			case Parasite::EShaderDataType::Bool:		return GL_BOOL;
 		}
 
 		PE_ASSERT(false, "Failed to find OpenGL type for shader data type.");
@@ -63,7 +63,8 @@ namespace Parasite
 				ShaderDataTypeToOpenGLBaseType(Element.Type),
 				Element.bNormalized ? GL_TRUE : GL_FALSE,
 				Layout.GetStride(),
-				(const void*)Element.Offset);
+				//(const void*)Element.Offset);
+				(const void*)(uintptr_t)Element.Offset);
 			Index++;
 		}
